fix(cli-sub-agent): sanitize subtree pin from plan bash-step env (#1741)
Round-6 review found plan_cmd_exec::spawn_bash started workflow bash steps with
the parent process environment fully inherited, then set a nested CSA_DEPTH. An
ambient/user-controlled CSA_MODEL_SPEC + CSA_FORCE_IGNORE_TIER_SETTING pair would
then be read by a nested `csa run` inside the step as a valid inherited subtree
pin -- spoofing a pin and silently disabling tier routing. The r5 typed channel
covered the executor spawn paths but not this non-executor (weave bash) spawn.

- plan_cmd_exec: apply_sanitized_subtree_pin(&mut Command) env_removes every
  SUBTREE_PIN_ENV_KEYS on the bash child, then re-applies CSA's pin ONLY via the
  typed inherited_subtree_model_pin() channel (Some only for a genuine pinned
  child: CSA_DEPTH>0 + valid spec + force-ignore marker). A legitimately
  propagated pin still cascades; ambient values can no longer masquerade.
- Sibling audit (every child spawn that inherits parent env AND sets/propagates
  CSA_DEPTH): spawn_bash was the sole gap. executor.rs / transport_cli.rs already
  strip via STRIPPED_ENV_VARS + typed channel (r5); pipeline_env feeds the
  executor which strips by construction; executor_codex_tmux replays the
  already-sanitized executor env; pipeline_gate runs only at depth 0 and never
  bumps depth; run_cmd_execute_post_exec_gate strips all CSA_* wholesale;
  csa-hooks never set/propagate CSA_DEPTH (out of class). No other change needed.
- plan_cmd_exec.rs split: extracted #[cfg(test)] mod tests to
  plan_cmd_exec_tests.rs (#[path]) to stay under the 8000-token monolith gate.

Tests (plan_cmd_exec_tests.rs): ambient full pin + CSA_DEPTH=0 (root spoof) -> all
3 keys removed, none re-applied; CSA_DEPTH=2 + valid pin (genuine child) -> all 3
cascade with correct values. 10 pre-existing plan tests still pass.

A by-construction follow-up (read-once + env::remove_var at process startup, so no
spawn path can leak the keys) is tracked in #1750. Addresses the round-6 finding.
